+++ src/job_stack.rs
@@ -1,8 +1,8 @@
 use {
     crate::*,
     anyhow::{
-        anyhow,
         Result,
+        anyhow,
     },
 };